omp_server/dev_requirement.txt
amqp==5.0.6
asgiref==3.4.1
aspy.yaml==1.3.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2020.6.20
cfgv==3.3.1
charset-normalizer==2.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.1.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==2.4.0
django-ipware==4.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.0.12
identify==2.2.14
idna==2.8
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.10.1
kombu==5.1.0
MarkupSafe==1.1.1
nodeenv==1.6.0
packaging==21.0
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
prompt-toolkit==3.0.20
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.5
python-crontab==2.5.1
python-dateutil==2.8.0
PyYAML==5.3.1
pymysql==1.0.2
pytz==2020.1
redis==3.5.3
requests==2.21.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.24.3
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
salt==3002.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
paramiko==2.7.2
scp==0.14.1
emoji==1.5.0
pycrypto==2.6.1
concurrent-log-handler==0.9.19
